created method called GenerateFile()
-created : private TransactionGenerateFileController generator = new TransactionGenerateFileController();
-created :private List<Account> accounts;

this method is used to generate a CSV that will list all the transactions for every account. It creates a
an array list of Transactions, then goes through a for loop that will get every transaction related to each account
in the List<Account> accounts arraylist. It then passes the transactions List through the GenerateFile method paramater, then names the file, then finally generates a CSV